DS-1064: Move SimulatedContext class into lambda package
