+++ xyz-util/src/main/java/com/here/xyz/util/service/aws/lambda/SimulatedContext.java
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package com.here.xyz.util.service.aws;
+package com.here.xyz.util.service.aws.lambda;
 
 import com.amazonaws.services.lambda.runtime.ClientContext;
 import com.amazonaws.services.lambda.runtime.CognitoIdentity;
